refactor: 로그인 애노테이션 이름 변경
