+++ src/main/java/roomescape/member/argumentResolver/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.common.argumentResolver;
+package roomescape.member.argumentResolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Login {
+public @interface LoginMember {
 }
